# api/openapi.yaml
openapi: 3.0.3
info:
  title: SurveyFlow API
  description: SurveyFlow backend API for creating, distributing, and analyzing surveys.
  version: 1.0.0
  contact:
    name: SurveyFlow
    email: api@surveyflow.io
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/auth/login/2fa:
    post:
      operationId: login2FA
      tags: [Auth]
      summary: Login with 2FA
      description: Verify a two-factor authentication code.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, code]
              properties:
                email:
                  type: string
                  format: email
                code:
                  type: string
      responses:
        "200":
          description: 2FA verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          description: Invalid code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/auth/refresh:
    post:
      operationId: refreshToken
      tags: [Auth]
      summary: Refresh token
      description: Exchange a refresh token for new access and refresh tokens.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [Auth]
      summary: Forgot password
      description: Request a password reset email. Always returns 200.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        "200":
          description: Reset email sent (always)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /api/v1/auth/reset-password:
    post:
      operationId: resetPassword
      tags: [Auth]
      summary: Reset password
      description: Reset password using a token from the forgot-password email.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, new_password]
              properties:
                token:
                  type: string
                new_password:
                  type: string
                  minLength: 8
                  maxLength: 128
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/me:
    get:
      operationId: getMe
      tags: [Users]
      summary: Get current user
      description: Return the authenticated user's profile.
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      operationId: updateMe
      tags: [Users]
      summary: Update profile
      description: Update the current user's name or avatar.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /api/v1/me/password:
    put:
      operationId: changePassword
      tags: [Users]
      summary: Change password
      description: Change the current user's password.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Current password is incorrect
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/v1/surveys/{surveyId}/analytics/questions/{questionId}:
    get:
      operationId: getQuestionStats
      tags: [Analytics]
      summary: Question stats
      description: Get statistics for a specific question.
      parameters:
        - $ref: "#/components/parameters/SurveyId"
        - name: questionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Question statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionStats"
  /api/v1/surveys/{surveyId}/analytics/cross-tab:
    get:
      operationId: getCrossTab
      tags: [Analytics]
      summary: Cross-tabulation
      description: Get cross-tabulation data between two questions.
      parameters:
        - $ref: "#/components/parameters/SurveyId"
        - name: row_question_id
          in: query
          required: true
          schema:
            type: string
        - name: col_question_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cross-tabulation data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrossTabResponse"
        "400":
          description: Missing query parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/surveys/{surveyId}/analytics/dropoff:
    get:
      operationId: getDropoff
      tags: [Analytics]
      summary: Dropoff rates
      description: Get question-by-question dropoff rates.
      parameters:
        - $ref: "#/components/parameters/SurveyId"
      responses:
        "200":
          description: Dropoff data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DropoffResponse"
  /api/v1/surveys/{surveyId}/analytics/ai-insights:
    post:
      operationId: getAIInsights
      tags: [Analytics]
      summary: AI insights
      description: Get AI-generated insights about survey responses.
      parameters:
        - $ref: "#/components/parameters/SurveyId"
      responses:
        "200":
          description: AI insights
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIInsightsResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SurveyId:
      name: surveyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
          example: BAD_REQUEST
        message:
          type: string
          example: invalid request body
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
...
    RefreshTokenRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    TokenResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            access_token:
              type: string
            refresh_token:
              type: string
            expires_at:
              type: string
              format: date-time
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        two_factor_enabled:
          type: boolean
        org_id:
          type: string
          format: uuid
        org_name:
          type: string
        role:
          type: string
          enum: [owner, admin, member, viewer]
        created_at:
          type: string
          format: date-time
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        avatar_url:
          type: string
          nullable: true
    ChangePasswordRequest:
      type: object
      required: [current_password, new_password]
      properties:
        current_password:
          type: string
        new_password:
          type: string
          minLength: 8
          maxLength: 128
...
    QuestionStats:
      type: object
      properties:
        question_id:
          type: string
        question_type:
          type: string
        question_title:
          type: string
        total_answers:
          type: integer
        skipped_count:
          type: integer
        choice_stats:
          type: array
          items:
            type: object
            properties:
              choice_id:
                type: string
              label:
                type: string
              count:
                type: integer
              percentage:
                type: number
        numeric_stats:
          type: object
          properties:
            mean:
              type: number
            median:
              type: number
            min:
              type: number
            max:
              type: number
            std_dev:
              type: number
        text_stats:
          type: object
          properties:
            total_words:
              type: integer
            avg_word_count:
              type: number
            top_keywords:
              type: array
              items:
                type: string
            sentiment_score:
              type: number
              nullable: true
    CrossTabResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            row_question_id:
              type: string
            column_question_id:
              type: string
            headers:
              type: array
              items:
                type: string
            rows:
              type: array
              items:
                type: object
                properties:
                  label:
                    type: string
                  values:
                    type: array
                    items:
                      type: integer
                  total:
                    type: integer
    DropoffResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            steps:
              type: array
              items:
                type: object
                properties:
                  step_id:
                    type: string
                  step_type:
                    type: string
                  step_label:
                    type: string
                  views:
                    type: integer
                  dropoffs:
                    type: integer
                  dropoff_rate:
                    type: number
    AIInsightsResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            summary:
              type: string
            key_findings:
              type: array
              items:
                type: string
            recommendations:
              type: array
              items:
                type: string
            sentiment:
              type: string
            themes:
              type: array
              items:
                type: object
                properties:
                  name:
                    type: string
                  description:
                    type: string
                  confidence:
                    type: number
                  mentions:
                    type: integer
            anomalies:
              type: array
              items:
                type: object
                properties:
                  type:
                    type: string
                  description:
                    type: string
                  severity:
                    type: string
                    enum: [low, medium, high]
                  question_id:
                    type: string
            generated_at:
              type: string
              format: date-time
...